refactor: extract ScanReviewForm from ScanEntryFlow + ratchet thresholds (#80, pt 2)
Completes the ScanEntryFlow decomposition. The review (post-OCR) and manual
entry steps shared most rendering — date chip, field rows, segmental table,
save button, date-picker overlay, the `showDatePicker`/`showOverwriteAlert`
state, and the `weightUnit` AppStorage — so both move into a single
ScanReviewForm view with a `mode: .review | .manual` switch at the body level.

The `@FocusState` for field focus stays on ScanEntryFlow so the gradient
background's tap-to-dismiss-focus still works; the form receives it as a
`FocusState<Bool>.Binding`. The dismiss action is threaded as `onClose`.
`reviewSaveButton` (a thin wrapper over `saveButton`) is dropped — review now
calls `saveButton` directly like manual.

  ScanEntryFlow.swift: 886 → 91 lines
  New ScanReviewForm.swift: 839 lines

With ScanEntryFlow no longer the pinning offender:

  - file_length:        1200 → 1000 (pinned by InBodyDocumentParser, 940)
  - type_body_length:   1000 → 950  (pinned by ScanReviewForm, <950)

Verified: make lint (0, --strict, with tightened thresholds), make build,
make test (272 logic + 13 UI incl. testManualScanEntrySaves end-to-end, all
green), and on-device manual entry (fields, date chip, focus dismissal,
save, dismiss the sheet).
